fixed health, added wallmaster movement

+++ Sprint_0/Collision/EnemyTileCollisionResponse.cs
@@ -12,7 +12,7 @@ public class EnemyTileCollisionResponse
          * Use sprite destination rectangles as hitboxes. 
          */
         Rectangle enemyRec = new Rectangle((int) enemy.xPos, (int) enemy.yPos, enemy.getWidth(), enemy.getHeight());
-        Rectangle tileRec = new Rectangle((int)tile.getXPos(), (int)tile.getYPos(), tile.getWidth(), tile.getHeight());
+        Rectangle tileRec = new Rectangle(tile.getXPos(), tile.getYPos(), tile.getWidth(), tile.getHeight());
 
 
         /*
